deleted plotting in turnExtreme; deleted Boat & sunspots
the plotting I deleted was in the TurnExtreme_Fat_Data.R script, and had been optional and set to FALSE for a long time – this was to look at the time series of extremes for all of the variables. I also hadn't been using the boat traffic or sunspot data, so I deleted those in the read-in file.