Finished fish across gear, added tax class
If a gear type was missing in a particular year, added NA's for all species that could have been caught by that gear type in that lake (based on whether that species was ever caught by that gear in that lake). Used gear that were used in >96% of years in the lake. Considered all VGN one gear type. Includes appropriate taxonomic classification.