Completed data read, taxon level
Completed my fixes to ReadFatData.R. This included a complete rework of the data read (saved manClean() indices, threw out columns that won't be used in analysis). Also included combining all time series in the 'long' direction. Finally, and most importantly, it included the taxonomic breakdown of the zooplankton and fish data, including a proper handling of fish cpue values in lake-years where a gear type was missing